refactor(*): Move IO utils to the utils directory
